+++ demos/multilevel/MultiLevelPlanningHyperCubeBenchmark.cpp
@@ -45,15 +45,10 @@
 #include <ompl/base/Path.h>
 #include <ompl/geometric/PathGeometric.h>
 
-// const double runtime_limit = 60;
-// const double memory_limit = 1024*1024; //in MB, but does not consider free operations from prev runs
-// const int run_count = 10;
-// unsigned int curDim = 50;
-
-const double runtime_limit = 2;
+const double runtime_limit = 60;
 const double memory_limit = 1024*20; //in MB, but does not consider free operations from prev runs
-const int run_count = 2;
-unsigned int curDim = 2;
+const int run_count = 10;
+unsigned int curDim = 100;
 
 int main(int argc, char **argv)
 {
@@ -111,7 +106,7 @@ int main(int argc, char **argv)
 
     addPlanner(benchmark, std::make_shared<og::SORRTstar>(si), range);
     addPlanner(benchmark, std::make_shared<og::SBL>(si), range);
-    addPlanner(benchmark, std::make_shared<og::SST>(si), range);
+    // addPlanner(benchmark, std::make_shared<og::SST>(si), range); //out of memory
     addPlanner(benchmark, std::make_shared<og::STRIDE>(si), range);
     addPlanner(benchmark, std::make_shared<og::FMT>(si), range);
 
@@ -125,7 +120,7 @@ int main(int argc, char **argv)
     addPlanner(benchmark, std::make_shared<og::KPIECE1>(si), range);
     addPlanner(benchmark, std::make_shared<og::BKPIECE1>(si), range);
     addPlanner(benchmark, std::make_shared<og::LBKPIECE1>(si), range);
-    addPlanner(benchmark, std::make_shared<og::PDST>(si), range);
+    // addPlanner(benchmark, std::make_shared<og::PDST>(si), range); //OOM
 
     addPlanner(benchmark, std::make_shared<og::PRM>(si), range);
     addPlanner(benchmark, std::make_shared<og::PRMstar>(si), range);